Develop 브랜치

+++ src/main/java/com/goorm/devlink/profileservice/controller/CalendarController.java
@@ -2,6 +2,7 @@ package com.goorm.devlink.profileservice.controller;
 
 import com.goorm.devlink.profileservice.dto.ScheduleDto;
 import com.goorm.devlink.profileservice.service.CalendarService;
+import com.goorm.devlink.profileservice.util.MessageUtil;
 import com.goorm.devlink.profileservice.vo.response.CalendarViewResponse;
 import com.goorm.devlink.profileservice.vo.request.ScheduleCreateRequest;
 import lombok.RequiredArgsConstructor;
@@ -9,17 +10,23 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequiredArgsConstructor
 public class CalendarController {
 
     private final CalendarService calendarService;
+    private final MessageUtil messageUtil;
 
     /** 유저 캘린더 조회 **/
     @GetMapping("/api/profile/schedule")
     public ResponseEntity<CalendarViewResponse> viewUserCalendar(@RequestParam("userUuid") String userUuid) {
+        if (userUuid.isEmpty()) {
+            throw new NoSuchElementException(messageUtil.getUserUuidEmptyMessage());
+        }
         List<ScheduleDto> scheduleDtos = calendarService.getCalendarScheduleDtos(userUuid);
         CalendarViewResponse calendarViewResponse = CalendarViewResponse.getInstanceForResponse(scheduleDtos);
         return new ResponseEntity<>(calendarViewResponse, HttpStatus.ACCEPTED);
@@ -27,7 +34,13 @@ public class CalendarController {
 
     /** 스케줄 생성 **/
     @PostMapping("/api/myprofile/schedule")
-    public ResponseEntity createCalendarSchedule(@RequestHeader("userUuid") String userUuid, @RequestBody ScheduleCreateRequest scheduleCreateRequest) {
+    public ResponseEntity createCalendarSchedule(@RequestHeader("userUuid") String userUuid, @Valid @RequestBody ScheduleCreateRequest scheduleCreateRequest) {
+        if (userUuid.isEmpty()) {
+            throw new NoSuchElementException(messageUtil.getUserUuidEmptyMessage());
+        }
+        if (scheduleCreateRequest.getMentoringUuid().isEmpty()) {
+            throw new NoSuchElementException(messageUtil.getMentoringUuidEmptyMessage());
+        }
         calendarService.saveCalendarByScheduleCreateRequest(userUuid, scheduleCreateRequest);
         return ResponseEntity.ok().build();
     }
@@ -35,6 +48,12 @@ public class CalendarController {
     /** 스케줄 취소 **/
     @DeleteMapping("/api/myprofile/schedule")
     public ResponseEntity cancelCalendarSchedule(@RequestHeader("userUuid") String userUuid, @RequestParam("mentoringUuid") String mentoringUuid) {
+        if (userUuid.isEmpty()) {
+            throw new NoSuchElementException(messageUtil.getUserUuidEmptyMessage());
+        }
+        if (mentoringUuid.isEmpty()) {
+            throw new NoSuchElementException(messageUtil.getMentoringUuidEmptyMessage());
+        }
         calendarService.deleteScheduleByUserUuidAndMentoringUuid(userUuid, mentoringUuid);
         return ResponseEntity.accepted().build();
     }
